requirements.txt
openai>=1.0.0
flask>=2.0.0
pymongo>=4.0.0
python-dotenv>=0.19.0
requests>=2.25.0